DebProject: Do not overwrite manifest.json
The deb platform backend included a manifest.json with pre-set
icons and start_url, which was copied to the application template.
This would overwrite the manifest.json holding all internal values
needed to keep app-tools working.

To fix this issue, manifest.json is no longer overwritten, but
the deb-specific values are set on the already existing manifest
in the create() method.

BUG=XWALK-4723